Build both shapeless 1.2.4 and shapless 2.
Make use of spaces feature of dbuild which allows us to build multiple
versions of the same project. We build shapeless 1.2.4 which is dependency
of spray. This allows us to get rid of the fork of spray.

We also build shapeless 2 which is dependency of specs2 and spire.

The specs2 and spire compile successfully. Spray is still failing with
compilations errors saved here:
https://gist.github.com/gkossakowski/10277880