Build both shapeless 1.2.4 and shapeless 2.
Make use of spaces feature of dbuild which allows us to build multiple
versions of the same project. We build both shapeless 1.2.4 and shapeless2
that are being published into separate spaces. Although spray upstream
depends on shapeless 1.2.4 I couldn't get it to compile so we keep our
fork of spray that is upgraded to shapeless 2.

Therefore shapeless 1.2.4 is a leaf in dependency graph: no other project
depends on shapeless 1.2.4.

https://gist.github.com/gkossakowski/10277880